requirements.txt
tqdm
numpy
kinpy
dotmap
pytorch-lightning
PyYAML
pytorch3d
omegaconf
loguru
scikit-image
scikit-learn
scipy
imageio
moviepy
colour
transforms3d==0.3.1
trimesh==3.9.5
opencv-python==4.5.1.48
pyrender==0.1.43
PyOpenGL==3.1.0
PyOpenGL_accelerate
matplotlib==3.2.2
joblib

git+https://github.com/cotton-ahn/human_body_prior
git+https://github.com/cotton-ahn/body_visualizer
git+https://github.com/MPI-IS/configer.git
git+https://github.com/MPI-IS/mesh.git
